requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
easyocr>=1.7.0
opencv-python>=4.8.0
pillow>=9.0.0
pyyaml>=6.0.0